moved definition of animation mappings to it's own file
